:truck: [#4980] Make few functions public API
to_multiple, rewrite_formio_components, and DataContainer need to be used in the JSON dump plugin
