+++ src/openforms/submissions/service.py
@@ -1,0 +1,3 @@
+from .logic.datastructures import DataContainer
+
+__all__ = ["DataContainer"]
